+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.8.0
+- Champions now equip ALEs 
+- ALE and gem chance now affect by luck
+
 ## 0.7.2
 - Fix Champions loot multiplying 4 times 
 - More ALE balancing